+++ examples/ruby/rideshare_rails/config/application.rb
@@ -36,15 +36,5 @@ module RideshareRails
     config.generators.system_tests = nil
     config.action_controller.include_all_helpers = true
     config.active_record.sqlite3_production_warning=false
-
-    Pyroscope.configure do |config|
-      config.app_name       = ENV.fetch("PYROSCOPE_APPLICATION_NAME", "rails-ride-sharing-app")
-      config.server_address = ENV.fetch("PYROSCOPE_SERVER_ADDRESS", "http://pyroscope:4040")
-      config.auth_token     = ENV.fetch("PYROSCOPE_AUTH_TOKEN", "")
-
-      config.tags = {
-        "region": ENV["REGION"] || "us-east",
-      }
-    end
   end
 end