+++ examples/tracing/jaeger/java-instrumentation/rideshare/src/main/java/rideshare/Main.java
@@ -1,0 +1,13 @@
+package rideshare;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Main.class, args);
+    }
+
+}