+++ examples/tracing/jaeger/java-instrumentation/rideshare/src/main/java/rideshare/car/CarService.java
@@ -1,0 +1,16 @@
+package rideshare.car;
+
+import rideshare.OrderService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CarService {
+    @Autowired
+    OrderService orderService;
+
+    public void orderCar(int searchRadius) {
+        orderService.findNearestVehicle(searchRadius, "car");
+    }
+
+}