+++ pkg/server/render.go
@@ -46,7 +46,8 @@ type renderMetadataResponse struct {
 	Query     string `json:"query"`
 	MaxNodes  int    `json:"maxNodes"`
 }
-type RenderResponse struct {
+
+type renderResponse struct {
 	flamebearer.FlamebearerProfile
 	Metadata renderMetadataResponse `json:"metadata"`
 }
@@ -149,22 +150,19 @@ func (rh *RenderHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 }
 
 // Enhance the flamebearer with a few additional fields the UI requires
-func (*RenderHandler) mountRenderResponse(flame flamebearer.FlamebearerProfile, appName string, gi *storage.GetInput, maxNodes int) RenderResponse {
+func (*RenderHandler) mountRenderResponse(flame flamebearer.FlamebearerProfile, appName string, gi *storage.GetInput, maxNodes int) renderResponse {
 	metadata := renderMetadataResponse{
-		flame.Metadata,
-		appName,
-		gi.StartTime.Unix(),
-		gi.EndTime.Unix(),
-		gi.Query.String(),
-		maxNodes,
+		FlamebearerMetadataV1: flame.Metadata,
+		AppName:               appName,
+		StartTime:             gi.StartTime.Unix(),
+		EndTime:               gi.EndTime.Unix(),
+		Query:                 gi.Query.String(),
+		MaxNodes:              maxNodes,
 	}
-
-	renderResponse := RenderResponse{
-		flame,
-		metadata,
+	return renderResponse{
+		FlamebearerProfile: flame,
+		Metadata:           metadata,
 	}
-
-	return renderResponse
 }
 
 func (rh *RenderHandler) renderParametersFromRequest(r *http.Request, p *renderParams) error {
@@ -205,23 +203,4 @@ func (rh *RenderHandler) renderParametersFromRequest(r *http.Request, p *renderP
 	p.format = v.Get("format")
 
 	return expectFormats(p.format)
-}
-
-func parseRenderRangeParams(r *http.Request, from, until string) (startTime, endTime time.Time, ok bool) {
-	switch r.Method {
-	case http.MethodGet:
-		fromStr, untilStr := r.URL.Query().Get(from), r.URL.Query().Get(until)
-		startTime, endTime = attime.Parse(fromStr), attime.Parse(untilStr)
-		return startTime, endTime, fromStr != "" || untilStr != ""
-	case http.MethodPost:
-		startTime, endTime = attime.Parse(from), attime.Parse(until)
-		return startTime, endTime, from != "" || until != ""
-	}
-
-	return time.Now(), time.Now(), false
-}
-
-type RenderTreeParams struct {
-	From  string `json:"from"`
-	Until string `json:"until"`
 }