+++ webapp/javascript/components/TimelineChart/TimelineChartWrapper.tsx
@@ -7,6 +7,9 @@ import type { Group } from '@pyroscope/models/src';
 import type { Timeline } from '@webapp/models/timeline';
 import { formatAsOBject } from '@webapp/util/formatDate';
 import Legend from '@webapp/pages/tagExplorer/components/Legend';
+import type { ExploreTooltipProps } from '@webapp/components/TimelineChart/ExploreTooltip';
+import type { ITooltipWrapperProps } from './TooltipWrapper';
+import TooltipWrapper from './TooltipWrapper';
 import TimelineChart from './TimelineChart';
 import styles from './TimelineChartWrapper.module.css';
 
@@ -74,6 +77,7 @@ type TimelineChartWrapperProps = TimelineDataProps & {
 
   /** selection type 'single' => gray selection, 'double' => color selection */
   selectionType: 'single' | 'double';
+  onHoverDisplayTooltip?: React.FC<ExploreTooltipProps>;
 };
 
 class TimelineChartWrapper extends React.Component<
@@ -131,6 +135,7 @@ class TimelineChartWrapper extends React.Component<
       },
       points: {
         show: false,
+        symbol: () => {}, // function that draw points on the chart
       },
       lines: {
         show: false,
@@ -152,6 +157,7 @@ class TimelineChartWrapper extends React.Component<
             ...flotOptions,
             lines: {
               show: true,
+              lineWidth: 0.8,
             },
             bars: {
               show: false,
@@ -280,6 +286,7 @@ class TimelineChartWrapper extends React.Component<
 
     const customFlotOptions = {
       ...flotOptions,
+      onHoverDisplayTooltip,
       xaxis: {
         ...flotOptions.xaxis,
         // In case there are few chunks left, then we'd like to add some margins to